backend/requirements.txt
asgiref==3.7.2
Django==3.2.3
djangorestframework==3.14.0
djoser==2.1.0
gunicorn==20.1.0
Pillow==10.0.1
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
